Add REGION_UNKNOWN/REGION_RESTRICTED defines; fix printf format for partialIndex
REGION_UNKNOWN (0) and REGION_RESTRICTED (255) are used by the firmware's transmission safety checks to skip TX in restricted/unknown areas. Also cast partialIndex to long and use %ld to match its type in the debug printf.

+++ src/h3lite.c
@@ -116,8 +116,8 @@ RegionId h3ToRegion(H3Index h3) {
     }
     
     // Debug output
-    printf("DEBUG h3ToRegion: h3=0x%llx baseCell=%d partialIndex=%d\n", 
-           (unsigned long long)h3, baseCell, partialIndex);
+    printf("DEBUG h3ToRegion: h3=0x%llx baseCell=%d partialIndex=%ld\n",
+           (unsigned long long)h3, baseCell, (long)partialIndex);
     
     // Search for matching region
     RegionId region = findRegion(baseCell, partialIndex);
